requirements.txt
snaql==0.4
click==6.6
PyYAML==3.11